Doctor Module Almost Completed

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { User, UserDocument } from './user.schema';
 import { Model } from 'mongoose';
@@ -6,10 +6,11 @@ import * as bcrypt from 'bcrypt';
 import { SignUpDto } from './dto/signUp.dto';
 import { PatientsService } from 'src/patients/patients.service';
 import { Patient } from 'src/patients/patient.schema';
+import { DoctorsService } from 'src/doctors/doctors.service';
 
 @Injectable()
 export class UsersService {
-    constructor(@InjectModel(User.name) private userModel: Model<UserDocument>,private patientservices:PatientsService){}
+    constructor(@InjectModel(User.name) private userModel: Model<UserDocument>,private patientservices:PatientsService, private doctorservices:DoctorsService){}
 
     async create(signupdto:SignUpDto): Promise<User>{
         const hashed = await bcrypt.hash(signupdto.password,10);
@@ -25,6 +26,17 @@ export class UsersService {
         user._id.toString()
     )
         }
+        if(user.role === 'doctor'){
+            await this.doctorservices.create({
+                name: signupdto.name,
+                age: signupdto.age,
+                gender: signupdto.gender,
+                contactNumber: signupdto.contactNumber,
+                department: signupdto.department
+            },
+        user._id.toString()
+    )
+        }
         return user
     }
 
@@ -32,10 +44,20 @@ export class UsersService {
         return await this.userModel.findOne({email});
     }
 
-    async removePatient(id:string): Promise<Patient>{
-        const patient = await this.patientservices.removePatientDetail(id)
+    async removePatient(id:string): Promise<any>{
+        const user = await this.userModel.findById(id);
+        if(!user)throw new NotFoundException('User does not Exist')
+
+        if(user.role === 'doctor'){
+            const doctor = await this.doctorservices.removeDoctorProfile(id)
+            await this.userModel.findByIdAndDelete(id)
+            return doctor
+        }
+
+        if(user.role === 'patient'){ const patient = await this.patientservices.removePatientDetail(id)
         await this.userModel.findByIdAndDelete(id)
         return patient
+        }
     }
 }
